requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.5
pandas==1.1.4
picamraw==1.2.64
Pillow==8.0.1
protobuf==3.14.0
psutil==5.8.0
py-cpuinfo==7.0.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==4.19.24
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
TBB==0.1
threadpoolctl==2.1.0
tifffile==2020.12.4

urllib3==1.26.2
py-cpuinfo==7.0.0
psutil==5.8.0
